requirements.txt
aniso8601==9.0.1
backcall==0.2.0
click==8.0.3
colorama==0.4.4
configparser==5.1.0
cvxopt==1.2.7
cycler==0.10.0
decorator==5.1.0
deprecation==2.1.0
Flask==2.0.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
graphviz==0.17
intervaltree==3.1.0
ipython==7.28.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.2
joblib==1.1.0
jsonpickle==2.0.0
kiwisolver==1.3.2
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mpmath==1.2.1
networkx==2.6.3
numpy==1.21.2
packaging==21.0
pandas==1.3.3
parso==0.8.2
pickleshare==0.7.5
Pillow==8.3.2
pm4py==2.2.14
prompt-toolkit==3.0.20
PuLP==2.1
pydotplus==2.0.2
Pygments==2.10.0
pygraphviz==1.7
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.3
pyvis==0.1.9
PyYAML==6.0
scikit-learn==1.0
scipy==1.7.1
six==1.16.0
sortedcontainers==2.4.0
StringDist==1.0.9
sympy==1.9
threadpoolctl==3.0.0
tqdm==4.62.3
traitlets==5.1.0
wcwidth==0.2.5
Werkzeug==2.0.2
